refactor: make minior improvements
